refactor(streaming): split harness dispatch from spawn manager
