+++ src/meridian/lib/streaming/spawn_manager.py
@@ -16,21 +16,17 @@ import psutil
 
 from meridian.lib.core.domain import SpawnStatus
 from meridian.lib.core.spawn_lifecycle import TERMINAL_SPAWN_STATUSES
-from meridian.lib.core.types import HarnessId, SpawnId, TransportId
+from meridian.lib.core.types import HarnessId, SpawnId
 from meridian.lib.harness import pi_lifecycle_events as pi_lifecycle
-from meridian.lib.harness.bundle import get_harness_bundle
 from meridian.lib.harness.connections.base import HarnessEvent
 from meridian.lib.harness.control_action import (
     ControlActionCoordinator,
     ControlActionType,
 )
-from meridian.lib.harness.errors import HarnessBinaryNotFound
-from meridian.lib.harness.permission_broker import PermissionBroker
 from meridian.lib.launch.launch_types import ResolvedLaunchSpec
 from meridian.lib.state import spawn_store
 from meridian.lib.state.atomic import append_text_line
 from meridian.lib.state.history import HarnessHistoryWriter
-from meridian.lib.state.paths import resolve_spawn_log_dir
 from meridian.lib.streaming.control_socket import ControlSocketServer
 from meridian.lib.streaming.drain_policy import (
     TURN_BOUNDARY_EVENT_TYPE,
@@ -53,6 +49,7 @@ from meridian.lib.streaming.heartbeat import heartbeat_loop
 from meridian.lib.streaming.pi_drain import PiDrainCoordinator
 from meridian.lib.streaming.pi_process_cleanup import terminate_pi_tracked_subspawns
 from meridian.lib.streaming.pi_subspawn_tracker import PiSubspawnTracker
+from meridian.lib.streaming.spawn_dispatch import dispatch_start
 from meridian.lib.streaming.types import InjectResult
 
 StartConnectionPort = Callable[
@@ -108,82 +105,6 @@ def _safe_connection_session_id(connection: object) -> str | None:
     except Exception:
         return None
     return session_id if isinstance(session_id, str) and session_id.strip() else None
-
-
-def _ensure_harness_bootstrap() -> None:
-    from meridian.lib.harness import ensure_bootstrap
-
-    ensure_bootstrap()
-
-
-async def dispatch_start(
-    config: ConnectionConfig,
-    spec: ResolvedLaunchSpec,
-) -> HarnessConnection[Any]:
-    """Dispatch one start call through bundle lookup and runtime type guard."""
-
-    from meridian.lib.harness.connections import get_connection_class
-
-    _ensure_harness_bootstrap()
-    bundle = get_harness_bundle(config.harness_id)
-    if not isinstance(spec, bundle.spec_cls):
-        raise TypeError(
-            f"HarnessBundle invariant violated: adapter for "
-            f"{bundle.harness_id} returned {type(spec).__name__}, "
-            f"expected {bundle.spec_cls.__name__}"
-        )
-
-    declared_transports = bundle.adapter.contract.transport.transport_ids
-    transport_id = _select_dispatch_transport(declared_transports)
-    connection_class = get_connection_class(config.harness_id, transport_id)
-    request_handler: PermissionBroker | None = None
-    connection_ref: dict[str, HarnessConnection[Any]] = {}
-
-    async def _runtime_event_sink(event: HarnessEvent) -> None:
-        await connection_ref["connection"].inject_runtime_event(event)
-
-    if config.harness_id is HarnessId.CODEX:
-        request_handler = PermissionBroker(
-            spawn_dir=resolve_spawn_log_dir(config.control_root, config.spawn_id),
-            event_sink=_runtime_event_sink,
-            auto_reject_runtime_requests=True,
-        )
-
-    connection: HarnessConnection[Any]
-    if request_handler is not None:
-        connection_factory = cast(
-            "Callable[..., HarnessConnection[Any]]",
-            connection_class,
-        )
-        try:
-            connection = connection_factory(request_handler=request_handler)
-        except TypeError:
-            connection = cast("Callable[[], HarnessConnection[Any]]", connection_class)()
-    else:
-        connection = cast("Callable[[], HarnessConnection[Any]]", connection_class)()
-
-    connection_ref["connection"] = connection
-
-    try:
-        await connection.start(config, spec)
-    except (FileNotFoundError, NotADirectoryError) as exc:
-        raise HarnessBinaryNotFound.from_os_error(
-            harness_id=config.harness_id,
-            error=exc,
-        ) from exc
-    return connection
-
-
-def _select_dispatch_transport(
-    declared_transports: tuple[TransportId, ...],
-) -> TransportId:
-    """Choose dispatch transport from adapter contract declarations."""
-
-    if len(declared_transports) == 1:
-        return declared_transports[0]
-    if TransportId.STREAMING in declared_transports:
-        return TransportId.STREAMING
-    return declared_transports[0]
 
 
 def _default_control_server_factory(
